refactor: split listicleItineraries types into model modules
types.ts was 838 lines and, despite the name, held 17 functions and 12
constants alongside the 30 types — angle pools, list tones, traveler
profile helpers, day resizing, moment guards.

Move the definitions into ./model, grouped by concern, and keep types.ts
as a barrel (838 -> 92 lines). It has to stay a barrel: ~110 modules
import from it.

Chose ./model over ./types/ deliberately — a types/ directory alongside
types.ts makes './types' ambiguous to resolve.

The Payload doc types were copied with sed rather than retyped, and
verified byte-identical. Their three near-identical repeated stop shapes
(whereStaying, items, and the deprecated top-level pair) are left
duplicated on purpose: collapsing them into shared aliases is a
type-level change, and they do differ (items adds moment/momentLabel, the
deprecated items also adds selectionReason). Worth a follow-up, not this
PR.

Verified: barrel's resolved export surface is identical (67/67 names,
compared via the TS checker against main); tsc error signature identical
to baseline; vitest 333 pass / 3 pre-existing failures, same as main;
check-boundaries output identical to main.
