+++ apps/ai-blog-writer/apps/frontend/src/features/listicleItineraries/model/common.ts
@@ -1,0 +1,8 @@
+import type { RelatedItemCollection } from './blockTypes'
+
+export type PayloadRichText = Record<string, unknown>
+
+export type PolymorphicRelatedItemValue = {
+  relationTo?: RelatedItemCollection | null
+  value?: number | { id?: number } | null
+}
